pyproject.toml
[project]
name = "rag-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.3",
    "chonkie[semantic]>=1.6.4",
    "chromadb>=1.5.8",
    "fastapi>=0.136.1",
    "fastmcp>=3.2.4",
    "litellm>=1.83.14",
    "mistralai>=1.12.4",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.27",
    "redis>=5",
    "scalar-fastapi>=1.8.2",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]
